Revert "Add lobby page and room management"
