+++ src/app/api/rooms/[roomId]/route.ts
@@ -1,16 +1,9 @@
 import { NextResponse } from 'next/server';
-import {
-  getParticipants,
-  saveParticipants,
-  deleteParticipants,
-  getRoom,
-  updateRoomName,
-} from '@/lib/db';
+import { getParticipants, saveParticipants, deleteParticipants } from '@/lib/db';
 
 export async function GET(req: Request, { params }: { params: { roomId: string } }) {
-  const room = getRoom(params.roomId);
-  if (!room) return NextResponse.json({ error: 'Room not found' }, { status: 404 });
-  return NextResponse.json({ name: room.name, participants: room.participants || [] });
+  const participants = getParticipants(params.roomId) || [];
+  return NextResponse.json({ participants });
 }
 
 export async function POST(req: Request, { params }: { params: { roomId: string } }) {
@@ -19,15 +12,6 @@ export async function POST(req: Request, { params }: { params: { roomId: string 
     return NextResponse.json({ error: 'Invalid data' }, { status: 400 });
   }
   saveParticipants(params.roomId, body.participants);
-  return NextResponse.json({ ok: true });
-}
-
-export async function PUT(req: Request, { params }: { params: { roomId: string } }) {
-  const body = await req.json().catch(() => ({}));
-  if (!body.name || typeof body.name !== 'string') {
-    return NextResponse.json({ error: 'Invalid data' }, { status: 400 });
-  }
-  updateRoomName(params.roomId, body.name);
   return NextResponse.json({ ok: true });
 }
