+++ src/app/page.tsx
@@ -1,92 +1,198 @@
 "use client";
 
-import { useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
+import { useState, useEffect, Suspense } from 'react';
+import { useRouter, useSearchParams } from 'next/navigation';
+import type { AvailabilityData } from '@/lib/types';
+import type { SuggestRestaurantInput, SuggestRestaurantOutput } from '@/ai/flows/suggest-restaurant';
+import { DatePollingForm } from '@/components/date-polling-form';
+import { AvailabilityMatrix } from '@/components/availability-matrix';
+import { RestaurantSuggestionForm } from '@/components/restaurant-suggestion-form';
+import { RestaurantResultCard } from '@/components/restaurant-result-card';
 import { Icons } from '@/components/icons';
+import { Button } from '@/components/ui/button';
+import { Loader2, RefreshCcw } from 'lucide-react';
+import { getRestaurantSuggestion } from './actions';
+import { useToast } from '@/hooks/use-toast';
 
-type Room = { id: string; name: string };
 
-export default function Lobby() {
+function HomeContent() {
   const router = useRouter();
-  const [rooms, setRooms] = useState<Room[]>([]);
-  const [newRoom, setNewRoom] = useState('');
+  const searchParams = useSearchParams();
 
-  const loadRooms = async () => {
-    try {
-      const res = await fetch('/api/rooms');
-      const data = await res.json();
-      setRooms(data.rooms || []);
-    } catch (e) {
-      console.error('Failed to load rooms', e);
-    }
-  };
+  const [roomId, setRoomId] = useState('');
+  const [availability, setAvailability] = useState<AvailabilityData | null>(null);
+  const [suggestions, setSuggestions] = useState<SuggestRestaurantOutput[] | null>(null);
+  const [bestDate, setBestDate] = useState<Date | null>(null);
+  const [excludedRestaurants, setExcludedRestaurants] = useState<string[]>([]);
+  const [searchCriteria, setSearchCriteria] = useState<SuggestRestaurantInput | null>(null);
+  const [isSearchingAgain, setIsSearchingAgain] = useState(false);
+  const { toast } = useToast();
 
   useEffect(() => {
-    loadRooms();
-  }, []);
+    let id = searchParams.get('room');
+    if (!id) {
+      id = crypto.randomUUID();
+      router.replace(`?room=${id}`);
+    }
+    setRoomId(id);
+  }, [router, searchParams]);
 
-  const handleCreate = async () => {
-    if (!newRoom.trim()) return;
-    try {
-      const res = await fetch('/api/rooms', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name: newRoom }),
+  const handleFindDates = (data: AvailabilityData) => {
+    setAvailability(data);
+    setSuggestions(null);
+    setBestDate(null);
+    setExcludedRestaurants([]);
+    setSearchCriteria(null);
+  };
+
+  const handleBestDateCalculated = (date: Date | null) => {
+    setBestDate(date);
+  };
+
+  const handleSuggestionGenerated = (data: SuggestRestaurantOutput[], input: SuggestRestaurantInput) => {
+    setSuggestions(data);
+    setSearchCriteria(input);
+    setExcludedRestaurants(prev => [...prev, ...data.map(s => s.restaurantName)]);
+  };
+
+  const handleSearchAgain = async () => {
+    if (!searchCriteria) return;
+
+    setIsSearchingAgain(true);
+    const result = await getRestaurantSuggestion({
+      ...searchCriteria,
+      excludedRestaurants,
+    });
+    setIsSearchingAgain(false);
+
+    if ('error' in result) {
+      toast({
+        variant: "destructive",
+        title: "Couldn't find more restaurants.",
+        description: result.error,
       });
-      const data = await res.json();
-      router.push(`/rooms/${data.id}`);
-    } catch (e) {
-      console.error('Failed to create room', e);
+    } else {
+      setSuggestions(result);
+      setExcludedRestaurants(prev => [...prev, ...result.map(s => s.restaurantName)]);
     }
   };
 
-  const handleRename = async (id: string, name: string) => {
+  const handleCopyLink = async () => {
     try {
-      await fetch(`/api/rooms/${id}`, {
-        method: 'PUT',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ name }),
+      await navigator.clipboard.writeText(window.location.href);
+      toast({
+        title: "Link copied!",
+        description: "Share it with your friends.",
       });
     } catch (e) {
-      console.error('Failed to rename room', e);
+      toast({
+        variant: "destructive",
+        title: "Copy failed",
+        description: "Unable to copy link.",
+      });
     }
+  };
+
+  const handleReset = async () => {
+    setAvailability(null);
+    setSuggestions(null);
+    setBestDate(null);
+    setExcludedRestaurants([]);
+    setSearchCriteria(null);
+    try {
+      if (roomId) {
+        await fetch(`/api/rooms/${roomId}`, { method: 'DELETE' });
+      }
+      router.refresh();
+    } catch (e) {
+      console.error('Failed to clear saved data', e);
+    }
+  }
+
+  const handleGoBack = () => {
+    setAvailability(null);
+    setSuggestions(null);
+    setBestDate(null);
+    setExcludedRestaurants([]);
+    setSearchCriteria(null);
   };
 
   return (
     <div className="flex min-h-screen w-full flex-col bg-background">
       <header className="sticky top-0 z-10 flex h-16 items-center gap-4 border-b bg-background/80 px-4 backdrop-blur-sm md:px-6">
-        <Icons.logo className="h-8 w-8 text-primary" />
-        <h1 className="font-headline text-2xl font-semibold text-foreground">GatherEase Lobby</h1>
+        <a href="/" className="flex items-center gap-4">
+          <Icons.logo className="h-8 w-8 text-primary" />
+          <h1 className="font-headline text-2xl font-semibold text-foreground">
+            GatherEase
+          </h1>
+        </a>
       </header>
       <main className="flex flex-1 flex-col items-center gap-8 p-4 md:p-8">
-        <div className="w-full max-w-xl space-y-6">
-          <div className="flex gap-2">
-            <Input value={newRoom} onChange={e => setNewRoom(e.target.value)} placeholder="New room name" />
-            <Button onClick={handleCreate}>Create</Button>
-          </div>
-          <ul className="space-y-2">
-            {rooms.map(room => (
-              <li key={room.id} className="flex items-center gap-2">
-                <Input
-                  className="flex-grow"
-                  value={room.name || ''}
-                  onChange={e => {
-                    const value = e.target.value;
-                    setRooms(prev => prev.map(r => r.id === room.id ? { ...r, name: value } : r));
-                  }}
-                  onBlur={e => handleRename(room.id, e.target.value)}
-                />
-                <Button variant="outline" onClick={() => router.push(`/rooms/${room.id}`)}>
-                  Open
+        <div className="w-full max-w-4xl space-y-8">
+          
+          {!availability && (
+            <div className="animate-in fade-in-0 slide-in-from-top-4 duration-500">
+              {roomId && <DatePollingForm onSubmit={handleFindDates} roomId={roomId} />}
+            </div>
+          )}
+
+          {availability && (
+            <div className="animate-in fade-in-0 slide-in-from-top-4 duration-500">
+              <AvailabilityMatrix
+                data={availability}
+                onBestDateCalculated={handleBestDateCalculated}
+                onReset={handleReset}
+                onGoBack={handleGoBack}
+              />
+            </div>
+          )}
+          
+          {bestDate && !suggestions && (
+            <div className="animate-in fade-in-0 slide-in-from-top-4 duration-500">
+              <RestaurantSuggestionForm onSuggestion={handleSuggestionGenerated} />
+            </div>
+          )}
+          
+          {suggestions && (
+            <div className="animate-in fade-in-0 slide-in-from-top-4 duration-500 space-y-8">
+              <div className="text-center">
+                 <h2 className="font-headline text-3xl font-bold">Your Suggested Spots!</h2>
+                 <p className="text-muted-foreground">Here are a few great options for your group, sorted by popularity.</p>
+              </div>
+              <div className="space-y-4">
+                {suggestions.map((s, index) => (
+                  <RestaurantResultCard key={s.restaurantName} data={s} rank={index + 1} />
+                ))}
+              </div>
+              <div className="flex justify-center gap-4">
+                 <Button variant="outline" onClick={handleSearchAgain} disabled={isSearchingAgain}>
+                  {isSearchingAgain ? (
+                    <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+                  ) : (
+                    <RefreshCcw className="mr-2 h-4 w-4" />
+                  )}
+                  Find New Suggestions
                 </Button>
-              </li>
-            ))}
-            {rooms.length === 0 && <p className="text-muted-foreground">No rooms yet.</p>}
-          </ul>
+                <Button variant="outline" onClick={handleCopyLink}>
+                  Share Link
+                </Button>
+              </div>
+            </div>
+          )}
         </div>
       </main>
+      <footer className="mt-auto py-4 text-center text-sm text-muted-foreground">
+        <p>Making plans with friends, simplified.</p>
+      </footer>
     </div>
   );
 }
+
+export default function Home() {
+  return (
+    <Suspense fallback={null}>
+      <HomeContent />
+    </Suspense>
+  );
+}
+
